+++ source/games/exhumed/src/ramses.cpp
@@ -159,7 +159,7 @@ void InitSpiritHead()
     nTalkTime = 1;
 }
 
-void DimSector(short nSector)
+void DimSector(int nSector)
 {
     short startwall = sector[nSector].wallptr;
     short nWalls = sector[nSector].wallnum;
